requirements-dev.txt
scipy
numpy
pandas
matplotlib
pint
codecov
pytest-cov
onshape_client